Crate citrus

This is the documentation for the citrus.

is an excitation and radiation transfer code that can be used to predict line and continuum radiation from an astronomical source. The code uses unstructured 3D Delaunay grids for photon transport and accelerated Lambda Iteration for population calculations.

For a detailed theoretical description of the code, please read the book.

Modules

collparts

config

constants

engine

grid

io

lines

macros

pops

raytrace

solver

source

tree

utils