citrus

Module config

Structs

Config

ContinuumImage

Domain

FreshRun

GridConfig

GridDensityMaxima

Represents a grid density maximum for use with the Modern sampling algorithm.

Image

ImageInput

The Image struct represents the configuration for an image to be generated.

LineImage

MolecularConfig

Orientation

Outputs

ParameterInput

This struct contains all basic settings such as number of grid points, model radius, input and output filenames, etc. Some of these parameters always need to be set by the user, while others are optional with preset default values. There is an exception to this rule, namely when restarting citrus with previously calculated populations. In that case, none of the non-optional parameters are required.

Parameters

RestartRun

SolverConfig

Spatial

Enums

DustConfig

ImageError

ImageKind

LineFreq

ParamError

RayTraceAlgorithm

RunMode

SamplingAlgorithm

Sampling algorithms to generate grid points for modeling.

SpectralAxis

Unit

Unit of the data in the image.

Functions

from_au 🔒

Deserialize a value in astronomical units (AU) to SI units (meters).

from_parsec 🔒

Deserialize a value in parsecs (pc) to SI units (meters).