citrus::config

Struct ParameterInput

pub struct ParameterInput {

    pub radius: f64,
    pub min_scale: f64,
    pub cmb_temp: f64,
    pub n_sink_points: usize,
    pub p_intensity: usize,
    pub enable_line_blending: bool,
    pub ray_trace_algorithm: RayTraceAlgorithm,
    pub sampling_algorithm: SamplingAlgorithm,
    pub lte_only: bool,
    pub init_lte: bool,
    pub polarization: bool,
    pub n_solve_iters: usize,
    pub output_file: Option<String>,
    pub binoutput_file: Option<String>,
    pub grid_file: Option<String>,
    pub dust_file: Option<String>,
    pub grid_in_file: Option<String>,
    pub reset_rng: bool,
    pub do_solve_rte: bool,
    pub n_mol_weights: Vec<f64>,
    pub dust_weights: Vec<f64>,
    pub collisional_partner_mol_weights: Vec<f64>,
    pub collisional_partner_ids: Vec<usize>,
    pub g_ir_data_files: Option<Vec<String>>,
    pub mol_data_files: Vec<String>,
    pub collisional_partner_names: Vec<String>,
    pub(crate) grid_out_files: Vec<String>,
    pub(crate) collisional_partner_user_set_flags: isize,
    pub(crate) radius_squ: f64,
    pub(crate) min_scale_squ: f64,
    pub(crate) taylor_cutoff: f64,
    pub(crate) grid_density_global_max: f64,
    pub(crate) ncell: usize,
    pub(crate) n_species: usize,
    pub(crate) n_densities: usize,
    pub(crate) n_line_images: usize,
    pub(crate) n_cont_images: usize,
    pub(crate) n_solve_iters_done: usize,
    pub(crate) do_interpolate_vels: bool,
    pub(crate) use_abun: bool,
    pub(crate) do_mol_calcs: bool,
    pub(crate) use_vel_func_in_raytrace: bool,
    pub(crate) edge_vels_available: bool,
    pub(crate) write_grid_at_stage: Vec<bool>,
    pub(crate) do_pre_grid: bool,

}

This struct contains all basic settings such as number of grid points, model radius, input and output filenames, etc. Some of these parameters always need to be set by the user, while others are optional with preset default values. There is an exception to this rule, namely when restarting citrus with previously calculated populations. In that case, none of the non-optional parameters are required.

Fields
radius: f64

This value sets the outer radius of the computational domain.

Note

It should be set large enough to cover the entire spatial extend of the model. In particular, if a cylindrical input model is used (e.g., the input file for the RATRAN code) one should not use the radius of the cylinder but rather the distance from the centre to the corner of the (r,z)-plane.

min_scale: f64

The smallest spatial scale sampled by the code. Structures smaller than minScale will not be sampled properly. If one uses spherical sampling (see below) this number can also be thought of as the inner edge of the grid. This number should not be set smaller than needed, because that will cause an undesirably large number of grid points to end up near the centre of the model.

cmb_temp: f64

Temperature of the cosmic microwave background in Kelvin.

It defaults to 2.725 K which is the value at zero redshift (i.e., the solar neighborhood). One should make sure to set this parameter properly when calculating models at a redshift larger than zero since $$ T_{\text{CMB}} = T_{\text{CMB},0} (1 + z) \text{K} $$ It should be noted that even though citrus can in this way take the change n CMB temperature with increasing $z$ into account, it does not (yet) take cosmological effects into account when ray tracing (such as stretching of frequencies when using Jansky as units).

n_sink_points: usize

The sinkPoints are grid points that are distributed randomly at [radius][Parameters::radius] forming the surface of the model. As a photon from within the model reaches a sink point it is said to escape and is not tracked any longer. The number of sink points is a user-defined quantity since the exact number may affect the resulting image as well as the running time of the code. One should choose a number that gives a surface density large enough not to cause artifacts in the image and low enough not to slow down the gridding too much. Since this is model dependent, a global best value cannot be given, but a useful range is between a few thousands and about ten thousand.

p_intensity: usize

This number is the number of model grid points. The more grid points that are used, the longer the code will take to run. Too few points however, will cause the model to be under-sampled with the risk of getting wrong results. Useful numbers are between a few thousands up to about one hundred thousand.

enable_line_blending: bool

If true, line blending is taken into account, however, only if there are any overlapping lines among the transitions found in the molecular data files. Enabling line blending will slow down the calculations considerably, in particular if there are multiple molecular data files.

By default, line blending is disabled.

ray_trace_algorithm: RayTraceAlgorithm

This parameter specifies the algorithm used by citrus to solve the radiative-transfer equations during ray-tracing. The default value of zero invokes the algorithm used in older versions a value of 1 invokes a new algorithm which is much more time-consuming but which produces much smoother images, free from step-artifacts.

If none of the four density-linked parameters are provided, citrus will attempt to guess the information, in a manner as close as possible to the way it was done in version 1.5 and earlier. This is safe enough when a single density value is returned, and only H2 provided as collision partner in the moldata file(s), but more complicated situations can very easily result in the code guessing wrongly. For this reason we encourage users to make use of these four parameters, although in order to preserve backward compatibility with old model.c files, we have not (yet) made them mandatory.

Note

There have been additional modifications to the raytracing algorithm which have significant effects on the output images since citrus-1.5. Image-plane interpolation is now employed in areas of the image where the grid point spacing is larger than the image pixel spacing. This leads both to a smoother image and a shorter processing time.

sampling_algorithm: SamplingAlgorithm

The sampling algorithm used for the model grid.

By default, the UniformExact is used which corresponds to default algorithm in newer versions of citrus.

lte_only: bool

If non-zero, citrus performs a direct LTE calculation rather than solving for the populations iteratively. This facility is useful for quick checks. The default is false, i.e., full non-LTE calculation.

init_lte: bool

If true, citrus solves for the level populations as usual, but LTE values are used to initialize the populations at the start instead of using values at T = 0.

polarization: bool

If true, citrus will calculate the polarized continuum emission. This parameter only has effect for continuum images. The resulting cube will have three channels containing the Stokes I, Q, and U of the continuum emission as according to the theory, the V component is zero. In order for polarization to work, a magnetic field must be defined.

When polarization is enabled, the implementation follows the DustPol code (Padovani et al. 2012), except that the expression used in Padovani et al. (2012) given for $\sigma^2$ has shown by Adele et al. (2015) to be small by a factor of 2. This correction has been implemented in citrus.

n_solve_iters: usize

This defines the number of solution iterations citrus should perform when solving non-LTE level populations. The default is currently 17. Note that it is now possible to run citrus in an incremental fashion. If the results of solving the RTE through N iterations are stored in a grid file via setting [Parameters::grid_out_files], then a second run of citrus, reading the grid file via [Parameters::grid_in_file], [Parameters::n_solve_iters] = M>N, will continue the RTE iterations starting at iteration N. (If you do this, your results will be slightly different, in a random way, than if you go to M iterations in one go, because the random seeds will be different.)

output_file: Option<String>

This is the file name of the output file that contains the level populations. If this parameter is not set, citrus will not output the populations. There is no default value.

binoutput_file: Option<String>

This is the file name of the output file that contains the grid, populations, and molecular data in binary format. This file is used to restart citrus with previously calculated populations. Once the populations have been calculated and the binoutputfile has been written, citrus can re-raytrace for a different set of image parameters without re-calculating the populations. There is no default value.

grid_file: Option<String>

This is the file name of the output file that contains the grid. If this parameter is not set, citrus will not output the grid. The grid file is written out as a VTK file. This is a formatted ascii file that can be read with a number of 3D visualizing tools (Visualization Tool Kit, Paraview, and others). There is no default value.

dust_file: Option<String>

Path to the dust opacity file.

This must be prpvided if any continuum images are requested. Optional if only line images are requested.

File Format

This table should be a two column ASCII file with wavelength in the first column and opacity in the second column. Currently citrus uses the same tables as RATRAN from Ossenkopf and Henning (1994), and so the wavelength should be given in microns (1e-6 meters) and the opacity in cm^2/g. This is the only place in citrus where SI units are not used. There is no default value. A future version of citrus may allow spatial variance of the dust opacities, so that opacities can be given as function of x, y, and z.

grid_in_file: Option<String>reset_rng: bool

If this is set non-zero, LIME will use the same random number seeds at the start of each solution iteration. This has the effect of choosing the same photon directions and frequencies for each iteration (although the directions and frequencies change randomly from one grid point to the next). This has the effect of decoupling any oscillation or wandering of the level populations as they relax towards convergence from the intrinsic Monte Carlo noise of the discrete solution algorithm. Best practice might involve alternating episodes with par->resetRNG =0 and 1, storing the intermediate populations via the :ref:I/O interface <grid-io>. Very little experience has been accumulated as yet with this facility.

do_solve_rte: bool

It is now possible to run LIME in two sessions: the first to solve the RTE and save the results to file, the second to read the file and create raytraced images from it. For a session of the first type you should set the number of images you specify via the :ref:img <images> parameter to zero, and give a value for one of the elements of :ref:par->gridOutFiles <grid-io>; for one of the second type you set :ref:par->gridInFile <grid-io> to the name of the file you just wrote, and include >0 image specifications in :ref:img <images>. There is a problem however for sessions of the first type: if you eventually want full-spectrum cubes then you will need some way to tell LIME to solve the RTE. In the past LIME has figured out if you want this from the presence of spectrum-type images in your :ref:img <images> list. To replace this capability we have added the present parameter. Thus, for first-stage sessions (supposing you choose to run LIME in that way rather than in the previous single-pass style) when you know that you will eventually want spectral cubes, you should set the present parameter. For all other cases it may be ignored.

n_mol_weights: Vec<f64>

We calculate the number density of each of its radiating species, at each grid point, by multiplying the abundance of the species (returned via the function of that name) by a weighted sum of the density values. This parameter allows the user to specify the weights in that sum.

dust_weights: Vec<f64>

Controls how the dust mass density and hence opacity is calculated. The calculation of dust mass density in older versions made use of a hard-wired average gas density value of 2.4, appropriate to a mix of 90% molecular hydrogen and 10% helium. This older formula will be used if none of the current four parameters are set.

collisional_partner_mol_weights: Vec<f64>collisional_partner_ids: Vec<usize>

The integer values are the codes given here. Currently recognized values range from 1 to 7 inclusive. E.g if the only colliding species of interest in your model is H2, your density function should return a single value, namely the density of molecular hydrogen, and (if you supply a collPartIds value at all) you should set collPartIds[0] = 1 (the LAMDA code for H2). However, if you use collisional partners that are not one of LAMDA partners, it is fine to use any of the values between 1 and 7 to match the density function with collisional information in the datafiles. Some of the messages in citrus will refer to the default LAMDA partner molecules, but this does not affect the calculations. In future we will introduce a better mechanism to allow the user to specify non-LAMDA collision partners.

This list acts as a link between the N density function returns (I’m using here N as shorthand for num_densities) and the M collision partner ID integers found in the moldatfiles. This allows us to associate density functions with the collision partner transition rates provided in the moldatfiles.

g_ir_data_files: Option<Vec<String>>mol_data_files: Vec<String>

This contains paths to molecular data files. It must be provided if any line images are specified (or [do_solve_rte][Parameters::do_solve_rte] is true). It is not read if only continuum images are requested.

Molecular data files contain the energy states, Einstein coefficients, and collisional rates which are needed by citrus to solve the excitation. These files must conform to the standard of the LAMDA database. If a data file name is give that cannot be found locally, citrus will try and download the file instead. When downloading data files, the filename can be give both with and without the extension .dat (i.e., CO or CO.dat). moldatfile is an array, so multiple data files can be used for a single citrus run. There is no default value.

collisional_partner_names: Vec<String>

Essentially this has only cosmetic importance since it has no effect on the functioning of citrus, only on the names of the collision partners which are printed to stdout. Its main purpose is to reassure the user who has provided transition rates for a non-LAMDA collision species in their moldatfile that they are actually getting these values and not some mysterious reversion to LAMDA.

grid_out_files: Vec<String>collisional_partner_user_set_flags: isizeradius_squ: f64min_scale_squ: f64taylor_cutoff: f64grid_density_global_max: f64ncell: usizen_species: usizen_densities: usizen_line_images: usizen_cont_images: usizen_solve_iters_done: usizedo_interpolate_vels: booluse_abun: booldo_mol_calcs: booluse_vel_func_in_raytrace: booledge_vels_available: boolwrite_grid_at_stage: Vec<bool>do_pre_grid: bool