citrus::config

Struct Parameters

pub struct Parameters {

    pub radius: f64,
    pub min_scale: f64,
    pub cmb_temp: f64,
    pub sink_points: usize,
    pub p_intensity: usize,
    pub blend: isize,
    pub ray_trace_algorithm: RayTraceAlgorithm,
    pub sampling_algorithm: isize,
    pub sampling: Sampling,
    pub lte_only: bool,
    pub init_lte: bool,
    pub polarization: bool,
    pub nthreads: usize,
    pub nsolve_iters: usize,
    pub output_file: String,
    pub binoutput_file: String,
    pub grid_file: String,
    pub pre_grid: String,
    pub restart: bool,
    pub dust: Option<String>,
    pub grid_in_file: String,
    pub reset_rng: bool,
    pub do_solve_rte: bool,
    pub nmol_weights: Vec<f64>,
    pub dust_weights: Vec<f64>,
    pub grid_density_max_locations: Vec<[f64; 3]>,
    pub grid_density_max_values: Vec<f64>,
    pub collisional_partner_mol_weights: Vec<f64>,
    pub collisional_partner_ids: Vec<usize>,
    pub grid_data_files: Option<Vec<String>>,
    pub mol_data_files: Vec<String>,
    pub collisional_partner_names: Vec<String>,
    pub grid_out_files: Vec<String>,
    pub collisional_partner_user_set_flags: isize,
    pub radius_squ: f64,
    pub min_scale_squ: f64,
    pub taylor_cutoff: f64,
    pub grid_density_global_max: f64,
    pub ncell: usize,
    pub n_images: usize,
    pub n_species: usize,
    pub num_densities: usize,
    pub do_pregrid: bool,
    pub num_grid_density_maxima: i32,
    pub num_dims: usize,
    pub n_line_images: usize,
    pub n_cont_images: usize,
    pub n_solve_iters_done: usize,
    pub do_interpolate_vels: bool,
    pub use_abun: bool,
    pub do_mol_calcs: bool,
    pub use_vel_func_in_raytrace: bool,
    pub edge_vels_available: bool,
    pub write_grid_at_stage: Vec<bool>,

}

Fields
radius: f64min_scale: f64cmb_temp: f64sink_points: usizep_intensity: usizeblend: isizeray_trace_algorithm: RayTraceAlgorithmsampling_algorithm: isizesampling: Samplinglte_only: boolinit_lte: boolpolarization: boolnthreads: usizensolve_iters: usizeoutput_file: Stringbinoutput_file: Stringgrid_file: Stringpre_grid: Stringrestart: booldust: Option<String>grid_in_file: Stringreset_rng: booldo_solve_rte: boolnmol_weights: Vec<f64>dust_weights: Vec<f64>grid_density_max_locations: Vec<[f64; 3]>grid_density_max_values: Vec<f64>collisional_partner_mol_weights: Vec<f64>collisional_partner_ids: Vec<usize>

This list acts as a link between the N density function returns (I’m using here N as shorthand for num_densities) and the M collision partner ID integers found in the moldatfiles. This allows us to associate density functions with the collision partner transition rates provided in the moldatfiles.

grid_data_files: Option<Vec<String>>mol_data_files: Vec<String>collisional_partner_names: Vec<String>

Essentially this has only cosmetic importance since it has no effect on the functioning of LIME, only on the names of the collision partners which are printed to stdout. Its main purpose is to reassure the user who has provided transition rates for a non-LAMDA collision species in their moldatfile that they are actually getting these values and not some mysterious reversion to LAMDA.

grid_out_files: Vec<String>collisional_partner_user_set_flags: isizeradius_squ: f64min_scale_squ: f64taylor_cutoff: f64grid_density_global_max: f64ncell: usizen_images: usizen_species: usizenum_densities: usizedo_pregrid: boolnum_grid_density_maxima: i32num_dims: usizen_line_images: usizen_cont_images: usizen_solve_iters_done: usizedo_interpolate_vels: booluse_abun: booldo_mol_calcs: booluse_vel_func_in_raytrace: booledge_vels_available: boolwrite_grid_at_stage: Vec<bool>