citrus

Module constants

Constants

CITRUS_GLOBAL_EPS

A small number to avoid division by zero Used to prevent numerical errors such as division by zero

CITRUS_RT_EPS

DENSITY_EXP

LOCAL_CMB_TEMP_SI

Local mean Cosmic Microwave Background (CMB) temperature in kelvin Temperature value from Fixsen (2009): 2.72548 K

NUM_OF_GRID_STAGES

N_DIMS

TREE_EXP

TYPICAL_ISM_DENS

Typical Interstellar Medium (ISM) density in particles per cubic centimeter A typical value for the ISM is 1000 particles/cm³