citrus

Module raytrace

Structs

BaryVelocityBuffer

ChainContext

Face

FaceBasis

GridInterp

InterCellKey 🔒

Intersect

This struct is meant to record all relevant information about the intersection between a ray (defined by a direction unit vector ‘dir’ and a starting position ‘r’) and a face of a simplex.

RTPreparation

RayData

Simplex

NOTE: it is assumed that vertex[i] is opposite the face that abuts with

Enums

Orientation

RTCError

RTCResult

Functions

assign_ray_on_image 🔒

calc_grid_cont_dust_opacity 🔒

calc_line_amp_erf 🔒

Approximates the average line-shape function along a path segment.

calc_line_amp_interp 🔒

The bulk velocity of the model material can vary significantly with position, thus so can the value of the line-shape function at a given frequency and direction. The present function calculates ‘vfac’, an approximate average of the line-shape function along a path of length ds in the direction of the line of sight.

calc_line_amp_sample 🔒

The bulk velocity of the model material can vary significantly with position, thus so can the value of the line-shape function at a given frequency and direction. The present function calculates ‘vfac’, an approximate average of the line-shape function along a path of length ds in the direction of the line of sight.

calc_second_order_shape_functions 🔒

calc_triangular_barycentric_coords 🔒

convert_cell_to_simplex 🔒

do_barycentric_interpolation 🔒

Performs linear interpolation over a simplex (triangle) using barycentric coordinates.

do_barycentric_interpolation_vel 🔒

do_barycentric_interpolations_vel 🔒

do_segment_interpolation 🔒

do_segment_interpolation_scalar 🔒

extract_grid_xs 🔒

follow_ray_through_cells 🔒

get_2d_cells 🔒

line_plane_intersect 🔒

Computes the distance to the next Voronoi face in a given direction.

locate_ray_on_image 🔒

prepare_raytrace 🔒

raytrace

trace_ray 🔒

Evaluates the light intensity along a line of sight for a given image pixel.

trace_ray_smooth 🔒

For a given image pixel position, this function evaluates the intensity of the total light emitted/absorbed along that line of sight through the (possibly rotated) model. The calculation is performed for several frequencies, one per channel of the output image.