citrus::raytrace

Function do_barycentric_interpolation

fn do_barycentric_interpolation(
    gip: &mut GridInterp,
    gp: &[Grid],
    mol_data: &[MolData],
    par: &Parameters,
    gis: &[usize; 3],
    intersect: &Intersect,
    x_cmpts_ray: RVector,
)

Performs linear interpolation over a simplex (triangle) using barycentric coordinates.

This function takes:

N values V_i at the vertices of a simplex (triangle),

The barycentric coordinates of a point within the simplex, and returns the interpolated value at that point.

This approach is equivalent to using linear shape functions in Finite Element Analysis. The interpolation uses shape functions Q_i, where each shape function:

Is zero at all vertices except the i-th,

Has value one at the i-th vertex.

The interpolated value at point r_ is computed as:

          N
          ┌
f(r_) =   │   V_i * Q_i(r_)
          └
         i=1

For linear interpolation, each shape function Q_i(r_) is equal to the i-th barycentric coordinate B_i of r_.

In this case, N == 3; the simplex is a triangular face of a Delaunay cell, and the interpolation point is the intersection of a ray with that face. This technique is used to interpolate several grid-based quantities.

For further reference, see the Wikipedia article on Barycentric coordinates.

Note

This is called from within the multi-threaded block.