citrus::raytrace

Function trace_ray_smooth

fn trace_ray_smooth(
    ray: &mut RayData,
    gips: &mut [GridInterp],
    gp: &[Grid],
    img: &Image,
    par: &Parameters,
    mol_data: &[MolData],
    rtp: &RTPreparation,
) -> Result<(), RTCError>

For a given image pixel position, this function evaluates the intensity of the total light emitted/absorbed along that line of sight through the (possibly rotated) model. The calculation is performed for several frequencies, one per channel of the output image.

Note that the algorithm employed here to solve the RTE is similar to that employed in the function calculateJBar() which calculates the average radiant flux impinging on a grid cell: namely the notional photon is started at the side of the model near the observer and ‘propagated’ in the receding direction until it ‘reaches’ the far side. This is rather non-physical in conception but it makes the calculation easier.

This version of traceray implements a new algorithm in which the population values are interpolated linearly from those at the vertices of the Delaunay cell which the working point falls within.

A note about the object ‘gips’: this is an array with 3 elements, each one a struct of type ‘gridInterp’. This struct is meant to store as many of the grid-point quantities (interpolated from the appropriate values at actual grid locations) as are necessary for solving the radiative transfer equations along the ray path. The first 2 entries give the values for the entry and exit points to a Delaunay cell, but which is which can change, and is indicated via the variables entryI and exitI (this is a convenience to avoid copying the values, since the values for the exit point of one cell are obviously just those for entry point of the next). The third entry stores values interpolated along the ray path within a cell.

Note

This is called from within the multi-threaded block.