citrus::source

Function stokes_angles

fn stokes_angles(
    mag_field: &RVector,
    rotation_matrix: &RMatrix,
) -> Result<RVector>

This function rotates the B-field vector from the model frame to the observer frame, then calculates and returns some useful values which will in function source_fn_polarized() make it easy to obtain the Stokes parameters of polarized submillimetre dust emission. (For an explanation of the reasons for choosing the particular quantities we do, see the comment in that function.)

Whenever one deals with polarized light, it is important to specify the coordinate systems carefully. In LIME the observer frame is defined such that, when the observer looks at the sky, the frame axes appear as follows:

           ^ Y
           |
           |
           |
    <------+
    X

The Z axis points into the page, away from the observer. Comparing this to normal astronomical coordinates one can see that X is in the direction of +ve right ascension and Y in the direction of +ve declination.

The IAU-recommended coordinate frame for expressing polarized light however is

           ^ X
           |
           |
           |
    <------O
    Y

with Z now emerging from the page (i.e pointing in the direction of propagation, towards the observer).

A vector defined in the LIME model basis can be converted to the observer basis by post-multiplying it with the image rotation matrix rotMat. (Another way of putting this is that the rows of rotMat are the unit vectors of the model coordinate frame expressed in the observer basis.) For the B field, this is expressed symbolically as

    Bp^T = B^T * rotation_matrix

where ^T denotes transpose.

Note

This is called from within a multi-threaded block.