citrus

Module utils

Modules

aux

Auxiliary functions

erf

Provides the error and related functions

interp

qrng

A simple quasi-random number generator based on the Halton sequence. The Halton sequence is a sequence of numbers that are generated by taking the fractional part of a number raised to a power. The Halton sequence is defined for any positive integer n, and the power is defined as 2^n.

Functions

calc_dust_data

calc_source_fn

The source function S is defined as j_nu/alpha, which is clearly not defined for alpha==0. However S is used in the algorithm only in the term (1-exp[-alpha*ds])*S, which is defined for all values of alpha. The present function calculates this term and returns it in the argument remnantSnu. For values of abs(alpha*ds) less than a pre- calculated cutoff supplied in ConfigInfo, a Taylor approximation is used. Note that the same cutoff condition holds for replacement of exp(-dTau) by its Taylor expansion to 3rd order.

gauss_line

get_dtg

get_dust_temp

get_erf

interpolate_kappa

This function:

planck_fn

progress_bar