citrus::utils

Module aux

Auxiliary functions

This module contains code from elsewhere which may be required by other functions in this module.

Evaluation of polynomials

Provides functions that don’t have a numerical solution and must be solved computationally (e.g. evaluation of a polynomial)

This module is directly copied from the statrs crate. The original code is licensed under the MIT license. It was purely done to avoid a dependency on the statrs crate.

The original code can be found here: https://github.com/statrs-dev/statrs/blob/master/src/function/evaluate.rs

Functions

polynomial

evaluates a polynomial at z where coeff are the coeffecients to a polynomial of order k where k is the length of coeff and the coeffecient to the kth power is the kth element in coeff. E.g. [3,-1,2] equates to 2z^2 - z + 3