citrus::utils

Module erf

Provides the error and related functions

This module is directly copied from the statrs crate. The original code is licensed under the MIT license. It was purely done to avoid a dependency on the statrs crate.

The original code can be found here: https://github.com/statrs-dev/statrs/blob/master/src/function/erf.rs

Constants

ERF_IMPL_AD 🔒

Polynomial coefficients for a denominator of erf_impl in the interval [1e-10, 0.5]

ERF_IMPL_AN 🔒

Polynomial coefficients for a numerator of erf_impl in the interval [1e-10, 0.5].

ERF_IMPL_BD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [0.5, 0.75].

ERF_IMPL_BN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [0.5, 0.75].

ERF_IMPL_CD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [0.75, 1.25].

ERF_IMPL_CN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [0.75, 1.25].

ERF_IMPL_DD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [1.25, 2.25].

ERF_IMPL_DN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [1.25, 2.25].

ERF_IMPL_ED 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [2.25, 3.5].

ERF_IMPL_EN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [2.25, 3.5].

ERF_IMPL_FD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [3.5, 5.25].

ERF_IMPL_FN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [3.5, 5.25].

ERF_IMPL_GD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [5.25, 8].

ERF_IMPL_GN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [5.25, 8].

ERF_IMPL_HD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [8, 11.5].

ERF_IMPL_HN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [8, 11.5].

ERF_IMPL_ID 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [11.5, 17].

ERF_IMPL_IN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [11.5, 17].

ERF_IMPL_JD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [17, 24].

ERF_IMPL_JN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [17, 24].

ERF_IMPL_KD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [24, 38].

ERF_IMPL_KN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [24, 38].

ERF_IMPL_LD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [38, 60].

ERF_IMPL_LN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [38, 60].

ERF_IMPL_MD 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [60, 85].

ERF_IMPL_MN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [60, 85].

ERF_IMPL_ND 🔒

Polynomial coefficients for a denominator in erf_impl in the interval [85, 110].

ERF_IMPL_NN 🔒

Polynomial coefficients for a numerator in erf_impl in the interval [85, 110].

ERF_INV_IMPL_AD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0, 0.5].

ERF_INV_IMPL_AN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0, 0.5].

ERF_INV_IMPL_BD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.5, 0.75].

ERF_INV_IMPL_BN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.5, 0.75].

ERF_INV_IMPL_CD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.75, 1] with x less than 3.

ERF_INV_IMPL_CN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.75, 1] with x less than 3.

ERF_INV_IMPL_DD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.75, 1] with x between 3 and 6.

ERF_INV_IMPL_DN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.75, 1] with x between 3 and 6.

ERF_INV_IMPL_ED 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.75, 1] with x between 6 and 18.

ERF_INV_IMPL_EN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.75, 1] with x between 6 and 18.

ERF_INV_IMPL_FD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.75, 1] with x between 18 and 44.

ERF_INV_IMPL_FN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.75, 1] with x between 18 and 44.

ERF_INV_IMPL_GD 🔒

Polynomial coefficients for a denominator of erf_inv_impl in the interval [0.75, 1] with x greater than 44.

ERF_INV_IMPL_GN 🔒

Polynomial coefficients for a numerator of erf_inv_impl in the interval [0.75, 1] with x greater than 44.

Functions

erf

erf calculates the error function at x.

erf_impl 🔒

erf_impl computes the error function at z. If inv is true, 1 - erf is calculated as opposed to erf

erf_inv

erf_inv calculates the inverse error function at x.

erf_inv_impl 🔒

erfc

erfc calculates the complementary error function at x.

erfc_inv

erfc_inv calculates the complementary inverse error function at x.