citrus::utils

Function calc_source_fn

pub fn calc_source_fn(dtau: f64, taylor_cutoff: f64) -> (f64, f64)

The source function S is defined as j_nu/alpha, which is clearly not defined for alpha==0. However S is used in the algorithm only in the term (1-exp[-alpha*ds])*S, which is defined for all values of alpha. The present function calculates this term and returns it in the argument remnantSnu. For values of abs(alpha*ds) less than a pre- calculated cutoff supplied in ConfigInfo, a Taylor approximation is used. Note that the same cutoff condition holds for replacement of exp(-dTau) by its Taylor expansion to 3rd order.