citrus

Module config

Structs

Config

Image

The Image struct represents the configuration for an image to be generated. For the field units, we have the following mapping:

Parameters

Enums

RayTraceAlgorithm

Sampling

Unit

Functions

load_config

parse_config

Parse the configuration file Arguments:

to_config 🔒

Type Aliases

Images 🔒

A container for all the images in the configuration file

MolDataVec 🔒