citrus

Module constants

Constants

AMU_CGS

Atomic mass unit in grams 1 AMU = 1.66053904 × 10⁻²⁷ kg = 1.66053904 × 10⁻²⁴ g

AMU_SI

Atomic mass unit (AMU) in kilograms 1 atomic mass unit = 1.66053904 × 10⁻²⁷ kg

ARCSEC_TO_RAD

Conversion factor: Arcseconds to radians 1 arcsec = π / 180 / 3600 radians

AU_CGS

Astronomical unit in centimeters 1 AU = 1.495978707 × 10¹¹ m = 1.495978707 × 10¹³ cm

AU_SI

Astronomical unit (AU) in meters 1 AU = 1.495978707 × 10¹¹ m

BOLTZMANN_CGS

Boltzmann constant in ergs per kelvin 1 erg/K = 1.38064852 × 10⁻²³ J/K × 10⁷

BOLTZMANN_SI

Boltzmann constant in joules per kelvin 1.38064852 × 10⁻²³ J/K

CITRUS_GLOBAL_EPS

A small number to avoid division by zero Used to prevent numerical errors such as division by zero

CITRUS_RT_EPS

GRAVITATIONAL_CONST_CGS

Gravitational constant in cubic centimeters per gram per second squared 6.67428 × 10⁻¹¹ m³/kg/s² = 6.67428 × 10⁻⁸ cm³/g/s²

GRAVITATIONAL_CONST_SI

Gravitational constant in cubic meters per kilogram per second squared 6.67428 × 10⁻¹¹ m³/kg/s²

HCKB

Constant: 100.0 * HPLANCK * CLIGHT / KBOLTZ Value: 1.43877735 (in appropriate units)

HPIP

Constant: HPLANCK * CLIGHT / (4.0 * PI * SPI) Value: 8.918502221e-27 (in appropriate units)

JULIAN_YEAR_SI

Length of the Julian year in seconds 365.25 days * 24 hours * 3600 seconds

LOCAL_CMB_TEMP_SI

Local mean Cosmic Microwave Background (CMB) temperature in kelvin Temperature value from Fixsen (2009): 2.72548 K

N_DIMS

PARSEC_SI

Parsecs in meters 1 parsec ≈ 3.08567758 × 10¹⁶ m

PLANCK_CGS

Planck constant in erg seconds 1 erg = 1 × 10⁻⁷ joules

PLANCK_SI

Planck constant in joule seconds 6.626070040 × 10⁻³⁴ J·s

SOLAR_MASS_CGS

Solar mass in grams 1 Solar mass = 1.9891 × 10³⁰ kg = 1.9891 × 10³³ g

SOLAR_MASS_SI

Solar mass in kilograms 1 Solar mass = 1.9891 × 10³⁰ kg

SOLAR_RADIUS_CGS

Solar radius in centimeters 1 Solar radius = 6.957 × 10⁸ m = 6.957 × 10¹⁰ cm

SOLAR_RADIUS_SI

Solar radius in meters 1 Solar radius = 6.957 × 10⁸ m

SPEED_OF_LIGHT_CGS

Speed of light in vacuum in centimeters per second 1 c = 2.99792458 × 10⁸ m/s = 2.99792458 × 10¹⁰ cm/s

SPEED_OF_LIGHT_SI

Speed of light in vacuum in meters per second 2.99792458 × 10⁸ m/s

STEFAN_BOLTZMANN_CGS

Stefan-Boltzmann constant in erg per square centimeter per kelvin to the fourth power per second 5.670367 × 10⁻⁸ W/m²/K⁴ = 5.670367 × 10⁻⁵ erg/cm²/K⁴/s

STEFAN_BOLTZMANN_SI

Stefan-Boltzmann constant in watts per square meter per kelvin to the fourth power 5.670367 × 10⁻⁸ W/m²/K⁴

TYPICAL_ISM_DENS

Typical Interstellar Medium (ISM) density in particles per cubic centimeter A typical value for the ISM is 1000 particles/cm³