citrus::grid

Function delaunay

pub fn delaunay(
    gp: &mut [Grid],
    num_points: usize,
    get_cells: bool,
    check_sink: bool,
) -> Result<DelaunayResult>

The principal purpose of this function is to perform a Delaunay triangulation for the set of points defined by the input argument ‘g’. This is achieved via routines in the 3rd-party package qhull.

A note about qhull nomenclature: a vertex is what you think it is - i.e., a point; but a facet means in this context a Delaunay triangle (in 2D) or tetrahedron (in 3D). This nomenclature arises because the Delaunay cells are indeed facets (or rather projections of facets) of the convex hull constructed in 1 higher dimension.

Required elements of structs: struct grid *gp: .id .x

Elements of structs are set as follows: struct grid *gp: .sink .numNeigh .neigh (this is malloc’d too large and at present not realloc’d.)

    cellType *dc (if getCells>0):
            .id
            .neigh
            .vertx